Playlist: add XSPF importer.
